refactor: organize files
폴더 구조 변경

역할에 맞게 스타일 파일 분리
