+++ src/components/layout/Footer.tsx
@@ -1,0 +1,5 @@
+import { NavBar } from './Footer.style';
+
+export default function Footer() {
+  return <NavBar></NavBar>;
+}
