+++ src/components/layout/Layout.style.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+// eslint-disable-next-line import/prefer-default-export
+export const ChildrenWrapper = styled.div`
+  padding-top: 3rem;
+  height: 100%;
+  width: 100%;
+`;
