+++ src/pages/Login/email.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/components/Layout';
+import Layout from '@/components/layout/Layout';
 
 import EmailLoginForm from './components/EmailLoginForm';
